requirements.txt
openai>=1.3.5
langgraph>=0.0.30
python-dotenv>=1.0.0